1、默认去除redis使用 2、默认去除日志的记录，可少配置一个mongodb或者 arangodb，
下载后导入数据库，并配置便可直接运行。

+++ src/com/lxyer/bbs/base/user/UserService.java
@@ -40,7 +40,7 @@ public class UserService extends BaseService {
         //log(user, 0, "用户登录成功.");
         UserInfo userInfo = user.createUserInfo();
 
-        sessions.setAsync(sessionExpireSeconds, loginBean.getSessionid(), userInfo.getUserid());
+        sessions.setAsync(sessionExpireSeconds, loginBean.getSessionid(), (long)userInfo.getUserid());
         retResult.setRetcode(0);
         retResult.setResult(userInfo);
         retResult.setRetinfo("登录成功.");
